CIV-9418 move exception to its own block
