+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/CreateClaimSdtController.java
@@ -3,16 +3,12 @@ package uk.gov.hmcts.reform.civil.controllers;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.civil.exceptions.Payload;
 import uk.gov.hmcts.reform.civil.model.CreateSDTResponse;
 import uk.gov.hmcts.reform.civil.modelsdt.CreateClaimSDT;
 import uk.gov.hmcts.reform.civil.service.CreateClaimFromSdtService;
@@ -34,12 +30,6 @@ public class CreateClaimSdtController {
     public ResponseEntity<CreateSDTResponse> createClaimSdt(@RequestBody CreateClaimSDT createClaimSDT) {
 
         return createClaimFromSdtService.buildResponse(createClaimSDT);
-    }
-
-    @PostMapping(path = "/exception/{testId}", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<CreateSDTResponse> welcome(@PathVariable("testId") int testId,
-                                                     @RequestBody @Valid Payload payload) {
-        return createClaimFromSdtService.buildException(testId);
     }
 }
