+++ src/main/java/uk/gov/hmcts/reform/civil/service/CreateClaimFromSdtService.java
@@ -5,15 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-import uk.gov.hmcts.reform.civil.exceptions.ApplicationException;
 import uk.gov.hmcts.reform.civil.mappings.CreateClaimCCD;
 import uk.gov.hmcts.reform.civil.mappings.CreateClaimMapper;
 import uk.gov.hmcts.reform.civil.model.CreateSDTResponse;
 import uk.gov.hmcts.reform.civil.modelsdt.CreateClaimSDT;
-
-import static uk.gov.hmcts.reform.civil.exceptions.ErrorDetails.INVALID_DATA;
-import static uk.gov.hmcts.reform.civil.exceptions.ErrorDetails.INVALID_DATA_CUSTOM;
 
 @Slf4j
 @Service
@@ -22,7 +17,6 @@ public class CreateClaimFromSdtService {
 
     private final CreateSDTResponse createSDTResponse;
     private final SubmitCreateClaim submitCreateClaim;
-    private final LocalValidatorFactoryBean validatorFactory;
 
     public ResponseEntity<CreateSDTResponse> buildResponse(CreateClaimSDT createClaimSDT) {
 
@@ -120,16 +114,4 @@ public class CreateClaimFromSdtService {
         return createClaimMapper.getCreateClaimCCD();
     }
 
-    public ResponseEntity<CreateSDTResponse> buildException(final int testId) {
-        //throw own error
-        if (testId == 1) {
-            throw new ApplicationException(INVALID_DATA, HttpStatus.BAD_REQUEST);
-        }
-        if (testId == 2) {
-            throw new ApplicationException(INVALID_DATA_CUSTOM, HttpStatus.BAD_REQUEST, "2333");
-        }
-        return new ResponseEntity<>(
-                CreateSDTResponse.builder().build(),
-                HttpStatus.CREATED);
-    }
 }
